combine reproductions into new filters e2e job
