--- e2e/test/scenarios/dashboard-filters/reproductions/20656-dashboard-breaks-for-user-without-card-permissions.cy.spec.js
@@ -1,83 +1,0 @@
-import { SAMPLE_DATABASE } from "e2e/support/cypress_sample_database";
-import { ADMIN_PERSONAL_COLLECTION_ID } from "e2e/support/cypress_sample_instance_data";
-import {
-  restore,
-  filterWidget,
-  visitDashboard,
-  editDashboard,
-  getDashboardCard,
-} from "e2e/support/helpers";
-
-const { PRODUCTS, PRODUCTS_ID } = SAMPLE_DATABASE;
-
-const filter = {
-  name: "ID",
-  slug: "id",
-  id: "11d79abe",
-  type: "id",
-  sectionId: "id",
-};
-
-const questionDetails = {
-  query: { "source-table": PRODUCTS_ID, limit: 2 },
-  // Admin's personal collection is always the first one (hence, the id 1)
-  collection_id: ADMIN_PERSONAL_COLLECTION_ID,
-};
-
-const dashboardDetails = {
-  parameters: [filter],
-};
-
-describe("issue 20656", () => {
-  beforeEach(() => {
-    restore();
-    cy.signInAsAdmin();
-  });
-
-  it("should allow a user to visit a dashboard even without a permission to see the dashboard card (metabase#20656, metabase#24536)", () => {
-    cy.createQuestionAndDashboard({ questionDetails, dashboardDetails }).then(
-      ({ body: { id, card_id, dashboard_id } }) => {
-        cy.request("PUT", `/api/dashboard/${dashboard_id}`, {
-          dashcards: [
-            {
-              id,
-              card_id,
-              row: 0,
-              col: 0,
-              size_x: 24,
-              size_y: 10,
-              parameter_mappings: [
-                {
-                  parameter_id: filter.id,
-                  card_id,
-                  target: ["dimension", ["field", PRODUCTS.ID, null]],
-                },
-              ],
-            },
-          ],
-        });
-
-        cy.signInAsNormalUser();
-
-        visitDashboard(dashboard_id);
-      },
-    );
-
-    // Make sure the filter widget is there
-    filterWidget();
-    // eslint-disable-next-line no-unscoped-text-selectors -- deprecated usage
-    cy.findByText("Sorry, you don't have permission to see this card.");
-
-    // Trying to edit the filter should not show mapping fields and shouldn't break frontend (metabase#24536)
-    editDashboard();
-
-    cy.findByTestId("edit-dashboard-parameters-widget-container")
-      .find(".Icon-gear")
-      .click();
-
-    getDashboardCard().within(() => {
-      cy.findByText("Column to filter on");
-      cy.icon("key");
-    });
-  });
-});
